third_party/WebKit/LayoutTests/svg/dom/svgpath-negative-pathLength-expected.txt

CONSOLE ERROR: line 4: Error: A negative value for path attribute <pathLength> is not allowed
